+++ home/common/apps.nix
@@ -18,5 +18,6 @@
     wps-office-cn
     keepassxc
     nextcloud-client          # Nextcloud 客户端
+    vscodium                  # VSCodium 编辑器（开源版 VS Code）
   ];
 }